requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
langchain==0.3.27
langchain-openai==0.3.28
langgraph==0.5.4
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
celery==5.5.3
bs4==0.0.2
azure-storage-blob==12.26.0
azure-identity==1.24.0
redis==6.4.0
uvicorn==0.35.0
async-property==0.2.2
channels==4.3.1
daphne==4.2.1
channels_redis==4.3.0